# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Resale AI System",
    "description": "Local-first resale automation pipeline",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/ebay/listing/{sku}": {
      "patch": {
        "tags": [
          "ebay"
        ],
        "summary": "Update Listing",
        "description": "Update a live eBay listing after publish.\nUses PUT /sell/inventory/v1/inventory_item/{sku} to update fields.\nAllowed updates: title, description, price, item_specifics.\nDoes NOT update: category_id, condition (these require relisting).\nReturns success or eBay error details.",
        "operationId": "update_listing_api_ebay_listing__sku__patch",
        "parameters": [
          {
            "name": "sku",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sku"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Updates"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/listings": {
      "get": {
        "tags": [
          "listings"
        ],
        "summary": "Get Listings",
        "description": "Return active listings (status = listed or exported).",
        "operationId": "get_listings_api_listings_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "all",
              "title": "Status"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/listings/sync": {
      "get": {
        "tags": [
          "listings"
        ],
        "summary": "Sync Listings",
        "description": "Sync active listings from eBay inventory API (paginated).",
        "operationId": "sync_listings_api_listings_sync_get",
        "parameters": [
          {
            "name": "skus",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Skus"
            }
          },
          {
            "name": "e2e_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "E2E Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/listings/push/{sku}": {
      "post": {
        "tags": [
          "listings"
        ],
        "summary": "Push To Ebay",
        "description": "Multi-step push of field updates to eBay.\nReturns per-step results so the UI can show progress.\nSteps: inventory_item, offer, promotion.",
        "operationId": "push_to_ebay_api_listings_push__sku__post",
        "parameters": [
          {
            "name": "sku",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sku"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/listings/end/{sku}": {
      "delete": {
        "tags": [
          "listings"
        ],
        "summary": "End Listing",
        "description": "Withdraw offer from eBay (ends listing, keeps inventory item).",
        "operationId": "end_listing_api_listings_end__sku__delete",
        "parameters": [
          {
            "name": "sku",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sku"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/listings/bulk/price": {
      "post": {
        "tags": [
          "listings"
        ],
        "summary": "Bulk Set Price",
        "operationId": "bulk_set_price_api_listings_bulk_price_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkPricePayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/listings/bulk/promo": {
      "post": {
        "tags": [
          "listings"
        ],
        "summary": "Bulk Set Promo",
        "operationId": "bulk_set_promo_api_listings_bulk_promo_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkPromoPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BulkPricePayload": {
        "properties": {
          "skus": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skus"
          },
          "price": {
            "type": "number",
            "title": "Price"
          }
        },
        "type": "object",
        "required": [
          "skus",
          "price"
        ],
        "title": "BulkPricePayload"
      },
      "BulkPromoPayload": {
        "properties": {
          "skus": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skus"
          },
          "promotion_pct": {
            "type": "number",
            "title": "Promotion Pct"
          }
        },
        "type": "object",
        "required": [
          "skus",
          "promotion_pct"
        ],
        "title": "BulkPromoPayload"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PushPayload": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "list_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "List Price"
          },
          "condition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Condition"
          },
          "condition_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Condition Notes"
          },
          "promotion_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Promotion Pct"
          },
          "promotion_enabled": {
            "type": "boolean",
            "title": "Promotion Enabled",
            "default": false
          },
          "photos_changed": {
            "type": "boolean",
            "title": "Photos Changed",
            "default": false
          }
        },
        "type": "object",
        "title": "PushPayload"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}